Add support for checking type declarations.
Parsed type declarations are translated in two passes. First a pre-environment
is generated, checking that all type constructor names are unique, which takes
all types as abstract. Then within this pre-environment, the real types are
determined, generalizing them to generic level. Afterwards, we check for
cyclic abbreviations and flag an error, as such abbreviations would cause
an endless loop for unify.